refactor: move user folder
