+++ src/app.module.ts
@@ -1,5 +1,4 @@
 import { CacheInterceptor, CacheModule, Module } from '@nestjs/common';
-import { AuthModule } from './auth/auth.module';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { factoriesModules } from './factories';
 import { APP_INTERCEPTOR } from '@nestjs/core';
@@ -8,7 +7,6 @@ import * as redisStore from 'cache-manager-redis-store';
 const configService = new ConfigService();
 @Module({
   imports: [
-    AuthModule,
     ...factoriesModules,
     ConfigModule.forRoot({ isGlobal: true }),
     CacheModule.register({
