+++ src/factories/auth/strategies/jwt.strategy.ts
