+++ src/factories/index.ts
@@ -1,5 +1,11 @@
+import { AuthModule } from './auth/auth.module';
 import { LikesModule } from './likes/likes.module';
 import { MatchesModule } from './matches/user.module';
 import { UserModule } from './user/user.module';
 
-export const factoriesModules = [LikesModule, UserModule, MatchesModule];
+export const factoriesModules = [
+  LikesModule,
+  UserModule,
+  MatchesModule,
+  AuthModule,
+];
