+++ src/factories/user/user.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { UserService } from './user.service';
-import { UserController } from './user.controller';
+import { UserService } from './services/user.service';
+import { UserController } from './controllers/user.controller';
 import { PrismaService } from 'src/core/recipes/prisma.service';
 
 @Module({
